SSR IOD of Beidou can be 10 bits

+++ c/include/rtcm3/messages.h
@@ -302,7 +302,7 @@ typedef struct {
 
 typedef struct {
   uint8_t sat_id;
-  uint8_t iode;
+  uint16_t iode;
   int32_t radial;
   int32_t along_track;
   int32_t cross_track;
